+++ src/main/api/user/request/CreateUserDTO.ts
@@ -1,0 +1,5 @@
+
+export interface CreateUserDTO {
+  "email": string;
+  "phone": string;
+}
